Add failsafe limit on queued output events
On the Gradle client, we are constrained to 64MB (by default). In
some cases, the number of output events in a throttling interval
can exceed the amount of memory available to the client.

This change adds a heuristic to render output events when too many
events are queued at one time.